+++ lib/project_template_cli.dart
@@ -1,0 +1,5 @@
+/// Project Template CLI Library.
+library project_template_cli;
+
+export 'project_template.dart';
+export 'src/project_template_cli.dart';